requirements.txt
matrix-io-proto==0.0.12
phue==1.0
protobuf==3.4.0
pyzmq==16.0.2
six==1.11.0
fuzzywuzzy==0.15.1
python-Levenshtein==0.12.0
pyalsaaudio==0.8.4
google-cloud-speech==0.30.0